2021.11

+++ src/generics_003/Main.java
@@ -3,7 +3,7 @@ package generics_003;
 class Main {
     public static void main(String[] args) {
        MyClass.Method(21);
-       MyClass.Method(2);
+       MyClass.Method((byte) 2);
        MyClass.Method(3.213);
        MyClass.Method(23.4f);
        MyClass.Method("123");
